Move bufferstuff from uart.c to new file bufferhandler.c

+++ system.c
@@ -32,7 +32,7 @@ void ConfigureOscillator(void)
  * Wait for a given number of milli-seconds using busy waiting scheme.
  * @param time - time in ms to wait.
  */
-void wait_ms(uint16_t time)
+void wait_ms(int time)
 {
     static long timel = 0;
     timel = time * 1000l;
